REFACT: server, location 따라 파일쪼개기
